requirements.txt
fastapi==0.115.0
uvicorn==0.30.0
openai-whisper
chromadb>=0.5.3
sentence-transformers==3.0.0
openai>=1.0.0
pdf2image==1.17.0
Pillow==10.4.0
python-dotenv==1.0.1
python-multipart==0.0.9
pydantic==2.9.0
